fixed remove in bchoc, add remove
